# file_scan.c: don't use calloc() in csum_whole_file()